Update titleize.rb
Adds in the fixes to this file little_words was being called beyond it's scope. I simply increased the scope of this variable with the $. Also the word & i variables inside of the titleized_words block were in the wrong order, meaning i was representing the word but being treated as the index leading to error.